refactor:AdminStoreList 페이지 기능, 뷰 분리
Page/Admin/Store/AdminStoreList
  - AdminStoreList.tsx
    - useGetMyStore
      - useMyStoresQuery와 결과를 저장하는 state와 함께 훅으로 묶었습니다.
    - AdminStoreListHeader
      - List 페이지 헤더 정보를 가지고 있습니다.
    - AdminStoreListItem
      - store 데이터를 가져와 리스트 UI로 출력합니다.
