+++ src/Page/Admin/Store/AdminStoreList/index.ts
@@ -1,0 +1,1 @@
+export { default as AdminStoreList } from "./AdminStoreList";
