+++ src/Page/Admin/Store/AdminUpdateStore.tsx
@@ -1,12 +1,8 @@
 import { useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
 
-import InputDefault from "@/Components/Form/InputDefault";
-import LabelDefault from "@/Components/Form/LabelDefault";
-import Loading from "@/Components/Loading/Loading";
-import useGetStore from "./hooks/useGetStore";
-import useLoadingComplete from "./hooks/useLoadingComplete";
-import useUpdateStore from "./hooks/useUpdateStore";
+import { InputDefault, LabelDefault, Loading } from "@/Components";
+import { useGetStore, useLoadingComplete, useUpdateStore } from "./hooks";
 import { IStoreFormProps } from "./interface";
 import { Form, InputContainer, StatusBar } from "./styles";
