+++ src/Page/Admin/Store/hooks/index.ts
@@ -1,3 +1,4 @@
 export { default as useAddStore } from "./useAddStore";
 export { default as useGetStore } from "./useGetStore";
 export { default as useUpdateStore } from "./useUpdateStore";
+export { default as useLoadingComplete } from "./useLoadingComplete";
